+++ heliotrope/adapters/endpoint/api/hitomi/image.py
@@ -5,7 +5,7 @@ from sanic_ext.extensions.openapi import openapi
 
 from heliotrope.application.exceptions import GalleryinfoNotFound
 from heliotrope.application.usecases.get.galleryinfo import GetGalleryinfoUseCase
-from heliotrope.infrastructure.sanic import HeliotropeRequest
+from heliotrope.infrastructure.sanic.app import HeliotropeRequest
 
 hitomi_image = Blueprint("hitomi_image", url_prefix="/image")
