+++ heliotrope/adapters/endpoint/api/hitomi/random.py
@@ -5,7 +5,7 @@ from sanic_ext.extensions.openapi import openapi
 from sanic_ext.extensions.openapi.types import Schema
 
 from heliotrope.application.usecases.get.info import GetRandomInfoUseCase
-from heliotrope.infrastructure.sanic import HeliotropeRequest
+from heliotrope.infrastructure.sanic.app import HeliotropeRequest
 
 hitomi_random = Blueprint("hitomi_random", url_prefix="/random")
