+++ heliotrope/infrastructure/sanic/app.py
@@ -6,13 +6,13 @@ from sanic.request import Request
 
 from heliotrope.application.javascript.interpreter import JavaScriptInterpreter
 from heliotrope.application.tasks.mirroring import MirroringTask
-from heliotrope.infrastructure.config import HeliotropeConfig
 from heliotrope.infrastructure.hitomila import HitomiLa
 from heliotrope.infrastructure.hitomila.repositories.galleryinfo import (
     HitomiLaGalleryinfoRepository,
 )
 from heliotrope.infrastructure.mongodb import MongoDB
 from heliotrope.infrastructure.mongodb.repositories.info import MongoDBInfoRepository
+from heliotrope.infrastructure.sanic.config import HeliotropeConfig
 from heliotrope.infrastructure.sqlalchemy import SQLAlchemy
 from heliotrope.infrastructure.sqlalchemy.repositories.galleryinfo import (
     SAGalleryinfoRepository,
